+++ src/main/java/org/liquidengine/legui/component/Layer.java
@@ -474,23 +474,15 @@ public class Layer<T extends Component> {
      * Used to add components.
      *
      * @param components components nodes to add.
-     *
-     * @return true if added.
-     *
-     * @see List#addAll(Collection)
      */
-    public boolean addAll(Collection<? extends Component> components) {
-        return container.addAll(components);
+    public void addAll(Collection<? extends Component> components) {
+        container.addAll(components);
     }
 
     /**
      * Used to remove component.
      *
      * @param component component to remove.
-     *
-     * @return true if removed.
-     *
-     * @see List#remove(Object)
      */
     public boolean remove(T component) {
         return container.remove(component);