+++ src/main/java/org/liquidengine/legui/component/event/checkbox/CheckBoxChangeValueEventListener.java
@@ -3,7 +3,7 @@ package org.liquidengine.legui.component.event.checkbox;
 import org.liquidengine.legui.listener.EventListener;
 
 /**
- * @author Aliaksandr_Shcherbin.
+ * @author ShchAlexander.
  */
 public interface CheckBoxChangeValueEventListener extends EventListener<CheckBoxChangeValueEvent> {