+++ src/main/java/org/liquidengine/legui/image/loader/ImageLoader.java
@@ -3,7 +3,7 @@ package org.liquidengine.legui.image.loader;
 import org.liquidengine.legui.image.LoadableImage;
 
 /**
- * Created by Aliaksandr_Shcherbin on 3/2/2017.
+ * Created by ShchAlexander on 3/2/2017.
  */
 public abstract class ImageLoader {