+++ src/main/java/org/liquidengine/legui/input/Mouse.java
@@ -14,7 +14,7 @@ import org.joml.Vector2f;
 /**
  * This class used to hold mouse state.
  * <p>
- * Created by Aliaksandr_Shcherbin on 2/9/2017.
+ * Created by ShchAlexander on 2/9/2017.
  */
 public class Mouse {
 
@@ -58,7 +58,7 @@ public class Mouse {
     }
 
     /**
-     * Created by Aliaksandr_Shcherbin on 2/9/2017.
+     * Created by ShchAlexander on 2/9/2017.
      */
     public enum MouseButton {
         /**