+++ src/main/java/org/liquidengine/legui/system/context/Context.java
@@ -14,7 +14,7 @@ import org.joml.Vector2i;
 import org.liquidengine.legui.component.Component;
 
 /**
- * Created by Aliaksandr_Shcherbin on 1/25/2017.
+ * Created by ShchAlexander on 1/25/2017.
  */
 public class Context {